+++ rust/sealevel/programs/hyperlane-sealevel-token-native/src/instruction.rs
@@ -19,7 +19,7 @@ pub fn init_instruction(
     let mut instruction = lib_init_instruction(program_id, payer, init)?;
 
     // Add additional account metas:
-    // 0. [writable] The native collateral PDA account.
+    // 0. `[writable]` The native collateral PDA account.
 
     let (native_collateral_key, _native_collatera_bump) =
         Pubkey::find_program_address(hyperlane_token_native_collateral_pda_seeds!(), &program_id);