+++ rust/utils/run-locally/src/cosmos/types.rs
@@ -1,6 +1,7 @@
 use std::{collections::BTreeMap, path::PathBuf};
 
 use hpl_interface::types::bech32_decode;
+use hyperlane_cosmos::RawCosmosAmount;
 
 use super::{cli::OsmosisCLI, CosmosNetwork};
 
@@ -35,12 +36,6 @@ pub struct TxResponse {
     pub logs: Vec<TxLog>,
 }
 
-#[derive(serde::Serialize, serde::Deserialize)]
-pub struct Coin {
-    pub denom: String,
-    pub amount: String,
-}
-
 #[derive(serde::Serialize, serde::Deserialize, Clone)]
 pub struct Codes {
     pub hpl_hook_merkle: u64,
@@ -73,7 +68,7 @@ pub struct Deployments {
 
 #[derive(serde::Serialize, serde::Deserialize)]
 pub struct BalanceResponse {
-    pub balances: Vec<Coin>,
+    pub balances: Vec<RawCosmosAmount>,
 }
 
 #[derive(serde::Serialize, serde::Deserialize)]
@@ -125,6 +120,7 @@ pub struct AgentConfig {
     pub prefix: String,
     pub signer: AgentConfigSigner,
     pub index: AgentConfigIndex,
+    pub gas_price: RawCosmosAmount,
 }
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
@@ -163,6 +159,10 @@ impl AgentConfig {
                 typ: "cosmosKey".to_string(),
                 key: format!("0x{}", hex::encode(validator.priv_key.to_bytes())),
                 prefix: "osmo".to_string(),
+            },
+            gas_price: RawCosmosAmount {
+                denom: "uosmo".to_string(),
+                amount: "0.05".to_string(),
             },
             index: AgentConfigIndex {
                 from: 1,