+++ solidity/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hyperlane-xyz/core
 
+## 3.4.0
+
+### Patch Changes
+
+- e06fe0b32: Supporting DefaultFallbackRoutingIsm through non-factory deployments
+- Updated dependencies [fd4fc1898]
+  - @hyperlane-xyz/utils@3.4.0
+
 ## 3.3.0
 
 ### Patch Changes